Update structure to match Tableaux
