+++ oriented_matroids/vector_oriented_matroid.py
@@ -19,13 +19,12 @@ AUTHORS:
 #                  http://www.gnu.org/licenses/
 ##############################################################################
 
-from sage.structure.unique_representation import UniqueRepresentation
-from sage.structure.parent import Parent
-from oriented_matroids import OrientedMatroids
+from oriented_matroids.abstract_oriented_matroid import AbstractOrientedMatroid
 from oriented_matroids.signed_vector_element import SignedVectorElement
+from sage.categories.sets_cat import Sets
 
 
-class VectorOrientedMatroid(UniqueRepresentation, Parent):
+class VectorOrientedMatroid(AbstractOrientedMatroid):
     r"""
     An oriented matroid implemented using vector axioms.
 
@@ -74,14 +73,13 @@ class VectorOrientedMatroid(UniqueRepresentation, Parent):
         - :class:`oriented_matroids.oriented_matroid.OrientedMatroid`
         - :class:`oriented_matroids.oriented_matroids_category.OrientedMatroids`
     """
-    Element = SignedVectorElement
-
     @staticmethod
-    def __classcall__(cls, data, groundset=None):
+    def __classcall__(cls, data, groundset=None, category=None):
         """
         Normalize arguments and set class.
         """
-        category = OrientedMatroids()
+        if category is None:
+            category = Sets()
         return super(VectorOrientedMatroid, cls) \
             .__classcall__(cls, data, groundset=groundset, category=category)
 
@@ -89,7 +87,7 @@ class VectorOrientedMatroid(UniqueRepresentation, Parent):
         """
         Initialize ``self``.
         """
-        Parent.__init__(self, category=category)
+        AbstractOrientedMatroid.__init__(self, category=category)
 
         # Set up our vectors
         vectors = []
